requirements.txt
Flask==3.0.3
pytest==8.3.3
pytest-mock==3.14.0
PyPDF2==3.0.1
Pillow==10.0.1
pdf2image==1.16.3
pytesseract==0.3.10
openai==1.51.0
python-docx==0.8.11
openpyxl==3.1.2
numpy==1.24.3
matplotlib==3.7.2
requests==2.31.0
python-dotenv==1.0.0
httpx==0.27.0

# Async processing and caching
celery[redis]==5.3.4
redis==5.0.1
aioredis==2.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Monitoring and metrics
psutil==5.9.6
prometheus-client==0.19.0

# Production server
gunicorn==21.2.0
gevent==23.9.1

# Additional utilities for scaling
kombu==5.3.4
billiard==4.2.0
vine==5.1.0
